requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-telegram-bot==22.7
python-dotenv==1.0.1
